requirements.txt
asyncpg
aiogram==2.7
aiohttp
python-dotenv
pillow
Django
psycopg2-binary
psycopg2
requests
bs4
pandas
openpyxl
xlsxwriter
xlrd
aioschedule
yookassa